requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-embed-video==1.4.10
django-livereload-server==0.5.1
django-unfold==0.46.0
idna==3.10
npm==0.1.1
optional-django==0.1.0
pillow==11.1.0
python-decouple==3.8
razorpay==1.4.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tornado==6.4.2
urllib3==2.3.0
